Накопительное обновление:
Исправил баг, из-за которого изображение в окне искажалось(растягивалось/сжималось в зависимости от отношения размеров окна)

Оптимизировал код модуля камеры и окна(теперь можно полноценно добавлять флаги окон)

Добавил функционал, связанный с отслеживанием vec2 (башня танка игрока следит за мышью, а боты - за танком игрока)

Мелкие исправления багов

+++ MyEngine/Camera2D.hpp
@@ -15,14 +15,14 @@ namespace MyEngine
 		glm::mat4 m_ortho_matrix;
 		float m_scale;
 		bool m_need_update;
-		int m_screen_width;
-		int m_screen_height;
+		float m_screen_width;
+		float m_screen_height;
 
 	public:
 		Camera2D();
 		~Camera2D();
 
-		void init_camera(int screen_width, int screen_height);
+		void init_camera(float screen_width, float screen_height);
 
 		void update();
